fixes + features
+ parse(code, argnames) - parse input string as function body with passed parameters.
XXX: Move parse to Function
~ Fixed BytecudeFunction call does not return value.

+++ include/ASTPrinter.h
@@ -18,7 +18,7 @@ void printAST(ck_ast::ASTNode *localroot) { // XXX: how to save line numbers?
 			break;
 		
 		case ck_token::INTEGER:
-			wprintf(L"%d", *(long long*) (localroot->objectlist->object));
+			wprintf(L"%d", *(int64_t*) (localroot->objectlist->object));
 			break;
 			
 		case ck_token::BOOLEAN:
