+++ include/objects/Native.h
@@ -41,7 +41,7 @@ namespace ck_objects {
 		static bool load(const std::wstring& filename, ck_vobject::vscope*, const std::vector<ck_vobject::vobject*>&);
 		
 		// Must return integer representation of an object
-		virtual long long int_value();
+		virtual int64_t int_value();
 		
 		// Must return string representation of an object
 		virtual std::wstring string_value();
